kinda big changes and not working but just to fix a point in time
