+++ caoslib/main.py
@@ -1,6 +1,4 @@
-from src.initialization import init, init_session
-from src.validator import validate_cookies, validate_session
-from src.summary import summary
+from src.initialization import init
 from src.sync import sync
 from src.standings import standings
 from src.style import style
@@ -18,35 +16,43 @@ import requests
 
 if __name__ == '__main__':
     args = arguments.Args()
-    if 'init' in args:
-        init()
-        puts(colored.green("Successfully logged in"))
-    elif 'init-session' in args:
-        init_session()
-        puts(colored.green("Successfully logged in"))
-
-    validate_cookies()
-
-    # recover previous session
-    session = requests.session()
-    with open(COOKIES_PATH, 'rb') as cookiesfile:
-        session.cookies.update(pickle.load(cookiesfile))
-
-    validate_session(session)
-
     flags = args.flags
-    if 'status' in args:
-        summary(session, '--solved' in flags or '-s' in flags)
-    elif 'sync' in args:
-        sync_samples = '--sync-samples' in flags or '-ssa' in flags or '--sync-all' in flags or '-sa' in flags
-        sync_statements = '--sync-statements' in flags or '-sst' in flags or '--sync-all' in flags or '-sa' in flags
-        sync(session, sync_samples, sync_statements)
-    elif 'standings' in args:
-        standings(session)
+    if 'init' in args:
+        init_project()
+        puts(colored.green("Init successful"))
     elif 'test' in args:
         test(args)
     elif 'style' in args:
         style(args)
+    else:
+        from src.web import init_session, login
+        from src.validator import validate_cookies, validate_session
+        from src.summary import summary
+
+        if 'login' in args:
+            login()
+            puts(colored.green("Successfully saved password and logged in"))
+            init_session()
+
+        #need edjude for extra data
+        validate_cookies()
+
+        # recover previous session
+        session = requests.session()
+        with open(COOKIES_PATH, 'rb') as cookiesfile:
+            session.cookies.update(pickle.load(cookiesfile))
+
+        if not validate_session(session):
+            init_session()
+
+        if 'status' in args:
+            summary(session, '--solved' in flags or '-s' in flags)
+        elif 'sync' in args:
+            sync_samples = '--sync-samples' in flags or '-ssa' in flags or '--sync-all' in flags or '-sa' in flags
+            sync_statements = '--sync-statements' in flags or '-sst' in flags or '--sync-all' in flags or '-sa' in flags
+            sync(session, sync_samples, sync_statements)
+        elif 'standings' in args:
+            standings(session)
 
     # menu = login(session)
     # print(menu)
