+++ caoslib/utils/constants.py
@@ -1,15 +1,7 @@
 import os
 
-#folder path
-def FOLDER_PATH():
-    with open(os.path.split(__file__)[0] + "/../files/path.txt", 'r') as folder_constant_file:
-        path = folder_constant_file.read()
-    return path
 
-# important paths
-CONFIG_PATH = FOLDER_PATH() + '/config.ini'
-COOKIES_PATH = FOLDER_PATH() + '/cookies.owo'
-LINKS_PATH = FOLDER_PATH() + '/links.json'
+# tasks data path
 CAOS_DIR = os.getenv('HOME') + '/programming/caos'
 
 # links
@@ -43,3 +35,15 @@ CLANG_FORMAT_STYLE_STRING = """\"{
     BreakBeforeBraces: Custom,
     BraceWrapping: { AfterEnum: true, AfterStruct: true }
 }\""""
+
+#lib path
+def LIB_PATH():
+    with open(os.path.split(__file__)[0] + "/../files/path.txt", 'r') as folder_constant_file:
+        path = folder_constant_file.read()
+    return path
+
+
+#relative to lib paths
+CONFIG_PATH = LIB_PATH() + '/config.ini'
+COOKIES_PATH = LIB_PATH() + '/cookies.owo'
+LINKS_PATH = LIB_PATH() + '/links.json'
